feat: #17 reduce code duplication in authentication components
