+++ src/views/Authentication/ResetPassword.tsx
@@ -1,32 +1,11 @@
-import { Container, Flex, Stack, useBreakpointValue } from '@chakra-ui/react';
-import { SyringeIllustartion } from './SyringeIllustartion';
-import React from 'react';
+import React, { FC } from 'react';
 import ResetPasswordForm from '../../components/authentication/resetPassword';
+import AuthenticationWrapper from './AuthenticationWrapper';
 
-export default function ResetPassword() {
-  const isMobile = useBreakpointValue({ base: true, md: false });
+const ResetPassword: FC = () => (
+  <AuthenticationWrapper>
+    <ResetPasswordForm />
+  </AuthenticationWrapper>
+);
 
-  return (
-    <Container
-      overflowY={'hidden'}
-      maxW={'10xl'}
-      h={'100vh'}
-      bg={'background.900'}
-      centerContent
-    >
-      {isMobile ? (
-        <ResetPasswordForm />
-      ) : (
-        <Stack isInline spacing={12} align='center' justify={'center'}>
-          <ResetPasswordForm />
-          <Flex w={'full'}>
-            <SyringeIllustartion
-              height={{ sm: '22rem', lg: '26rem' }}
-              maxW={'40vw'}
-            />
-          </Flex>
-        </Stack>
-      )}
-    </Container>
-  );
-}
+export default ResetPassword;
