+++ src/views/Authentication/SignIn.tsx
@@ -1,32 +1,11 @@
-import { Container, Flex, Stack, useBreakpointValue } from '@chakra-ui/react';
-import { SyringeIllustartion } from '../Authentication/SyringeIllustartion';
 import SignInCard from '../../components/authentication/signinCard';
-import React from 'react';
+import React, { FC } from 'react';
+import AuthenticationWrapper from './AuthenticationWrapper';
 
-export default function SignIn() {
-  const isMobile = useBreakpointValue({ base: true, md: false });
+const SignIn: FC = () => (
+  <AuthenticationWrapper>
+    <SignInCard />
+  </AuthenticationWrapper>
+);
 
-  return (
-    <Container
-      overflowY={'hidden'}
-      maxW={'10xl'}
-      h={'100vh'}
-      bg={'background.900'}
-      centerContent
-    >
-      {isMobile ? (
-        <SignInCard />
-      ) : (
-        <Stack isInline spacing={12} align='center' justify={'center'}>
-          <SignInCard />
-          <Flex w={'full'}>
-            <SyringeIllustartion
-              height={{ sm: '22rem', lg: '26rem' }}
-              maxW={'40vw'}
-            />
-          </Flex>
-        </Stack>
-      )}
-    </Container>
-  );
-}
+export default SignIn;
