+++ src/views/Authentication/SignUp.tsx
@@ -1,32 +1,11 @@
-import { Container, Flex, Stack, useBreakpointValue } from '@chakra-ui/react';
-import { SyringeIllustartion } from './SyringeIllustartion';
-import React from 'react';
+import React, { FC } from 'react';
 import SignUpCard from '../../components/authentication/signupCard';
+import AuthenticationWrapper from './AuthenticationWrapper';
 
-export default function SignUp() {
-  const isMobile = useBreakpointValue({ base: true, md: false });
+const SignUp: FC = () => (
+  <AuthenticationWrapper>
+    <SignUpCard />
+  </AuthenticationWrapper>
+);
 
-  return (
-    <Container
-      overflowY={'hidden'}
-      maxW={'10xl'}
-      h={'100vh'}
-      bg={'background.900'}
-      centerContent
-    >
-      {isMobile ? (
-        <SignUpCard />
-      ) : (
-        <Stack isInline spacing={12} align='center' justify={'center'}>
-          <SignUpCard />
-          <Flex w={'full'}>
-            <SyringeIllustartion
-              height={{ sm: '22rem', lg: '26rem' }}
-              maxW={'40vw'}
-            />
-          </Flex>
-        </Stack>
-      )}
-    </Container>
-  );
-}
+export default SignUp;
